Remove atlas private client
